Updated 22\3\21

+++ application/Model/Core/Table.php
@@ -140,9 +140,11 @@
             return $this->getAdapter()->insert($query);
         }
     
-        public function delete(){
-            $id = $_GET['id'];
-            $query="Delete FROM `{$this->getTableName()}` WHERE  `{$this->getPrimaryKey()}` = '{$id}'";
+        public function delete($query = null){
+            if(!$query){
+                $id = $this->getRequest()->getGet('id');
+                $query="Delete FROM `{$this->getTableName()}` WHERE  `{$this->getPrimaryKey()}` = '{$id}'";
+            }
             return $this->getAdapter()->delete($query);  
         }
     }
